requirements.txt
flask == 3.1.0
tensorflow-cpu == 2.19.0
numpy == 2.1.3